CriptoLladosApp

+++ DAW/Criptoapp/CriptoLladosApp/criptoapp/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Criptomoneda(models.Model):
     nombre = models.CharField(max_length=100, unique=True)
-    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    precio = models.FloatField()
     fecha_creacion = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
